Add shared level overview and user profile screen
